Compare

+++ App.Core.ApplicationService/Dtos/MovieDtos/MovieCompareOutputDto.cs
@@ -7,7 +7,7 @@ namespace App.Core.ApplicationService.Dtos.MovieDtos
    public class MovieCompareOutputDto
     {
         public string[] Genres { get; set; }
-        public float RateByUsers { get; set; }
+        public double RateByUsers { get; set; }
         public int VisitCounts { get; set; }
     }
 }
